backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
Authlib==1.6.9
bandit==1.9.4
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
dparse==0.6.4
email-validator==2.3.0
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mdurl==0.1.2
nltk==3.9.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-magic==0.4.27
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
stevedore==5.7.0
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
PyJWT==2.10.1
slowapi==0.1.9
